Updated CONFIG file with more options. New features in HiPACE++ API and stage. Refactored many features into Source parent class, added features. Small fixes in Beam, Linac, Interstage and Spectrometer. Updated examples.

+++ abel/apis/elegant/elegant_api.py
@@ -4,24 +4,31 @@ from abel import CONFIG, Beam
 import scipy.constants as SI
 from abel.utilities.relativity import gamma2energy, energy2gamma
 
-def elegant_read_beam(filename):
+def elegant_read_beam(filename, tmpfolder=None):
     
     # create temporary stream file and folder
-    tmpfolder = CONFIG.temp_path + str(uuid.uuid4())
-    os.mkdir(tmpfolder)
-    tmpfile =  tmpfolder + '/stream.tmp'
+    make_new_tmpfolder = tmpfolder is None
+    if make_new_tmpfolder:
+        tmpfolder = CONFIG.temp_path + str(uuid.uuid4())
+        os.mkdir(tmpfolder)
+    tmpfile =  tmpfolder + '/stream_' + str(uuid.uuid4()) + '.tmp'
     
     # convert SDDS file to CSV file
-    subprocess.call(CONFIG.elegant_exec + 'sdds2stream ' + filename + ' -columns=x,xp,y,yp,t,p,dt > ' + tmpfile, shell=True)
+    subprocess.run(CONFIG.elegant_exec + 'sdds2stream ' + filename + ' -columns=x,xp,y,yp,t,p,dt > ' + tmpfile, shell=True)
     
     # extract charge
-    Q = float(subprocess.check_output([CONFIG.elegant_exec + 'sdds2stream ' + filename + ' -parameter=Charge'], shell=True))
+    try:
+        Q = float(subprocess.check_output([CONFIG.elegant_exec + 'sdds2stream ' + filename + ' -parameter=Charge'], shell=True))
+    except:
+        return None
     
     # load phasespace from CSV file
     phasespace = np.loadtxt(open(tmpfile, "rb"), delimiter=' ')
     
     # delete CSV file and temporary folder
-    shutil.rmtree(tmpfolder)
+    os.remove(tmpfile)
+    if make_new_tmpfolder:
+        shutil.rmtree(tmpfolder)
     
     # make beam (note: z is flipped)
     beam = Beam()
@@ -31,18 +38,20 @@ def elegant_read_beam(filename):
                        xps=phasespace[:,1], 
                        yps=phasespace[:,3], 
                        Es=gamma2energy(phasespace[:,5]),
-                       Q=Q)
+                       Q=Q) # TODO: deal with non-uniform weight particles
     beam.location = np.mean(phasespace[:,4])*SI.c
     
     return beam
 
 
-def elegant_write_beam(beam, filename):
+def elegant_write_beam(beam, filename, tmpfolder=None):
     
     # create temporary CSV file and folder
-    tmpfolder = CONFIG.temp_path + str(uuid.uuid4()) + '/'
-    os.mkdir(tmpfolder)
-    tmpfile = tmpfolder + 'beam.csv'
+    make_new_tmpfolder = tmpfolder is None
+    if make_new_tmpfolder:
+        tmpfolder = CONFIG.temp_path + str(uuid.uuid4()) + '/'
+        os.mkdir(tmpfolder)
+    tmpfile = tmpfolder + 'beam_' + str(uuid.uuid4()) + '.csv'
     
     # write beam phasespace to CSV (note: z/t is flipped)
     M = np.matrix([beam.xs(),beam.xps(),beam.ys(),beam.yps(),-1*beam.ts(),beam.gammas(),beam.ts()])
@@ -65,46 +74,38 @@ def elegant_write_beam(beam, filename):
     with open(filename, 'r') as f:
         lines = f.readlines()
         
-    line1 = 2
-    lines.insert(line1, '&parameter name=SVNVersion, description="SVN version number", type=string, &end\n')
-    lines.insert(line1, '&parameter name=IDSlotsPerBunch, description="Number of particle ID slots reserved to a bunch", type=long, &end\n')
-    lines.insert(line1, '&parameter name=Particles, description="Number of particles before sampling", type=long, &end\n')
-    lines.insert(line1, '&parameter name=Charge, units=C, description="Bunch charge before sampling", type=double, &end\n')
-    lines.insert(line1, '&parameter name=pCentral, symbol="p$bcen$n", units="m$be$nc", description="Reference beta*gamma", type=double, &end\n')
-    lines.insert(line1, '&parameter name=Step, description="Simulation step", type=long, &end\n')
-
-    line2 = 18
-    lines.insert(line2, '28584M'+'\n')
-    lines.insert(line2, str(int(len(beam)))+'\n')
-    lines.insert(line2, str(int(len(beam)))+'\n')
-    lines.insert(line2, ' '+str(beam.charge())+'\n')
-    lines.insert(line2, ' '+str(energy2gamma(beam.energy()))+'\n')
-    lines.insert(line2, str(1)+'\n')
-                 
-    with open(filename, 'w') as f:
-        f.write("".join(lines))
+        line1 = 2
+        lines.insert(line1, '&parameter name=SVNVersion, description="SVN version number", type=string, &end\n')
+        lines.insert(line1, '&parameter name=IDSlotsPerBunch, description="Number of particle ID slots reserved to a bunch", type=long, &end\n')
+        lines.insert(line1, '&parameter name=Particles, description="Number of particles before sampling", type=long, &end\n')
+        lines.insert(line1, '&parameter name=Charge, units=C, description="Bunch charge before sampling", type=double, &end\n')
+        lines.insert(line1, '&parameter name=pCentral, symbol="p$bcen$n", units="m$be$nc", description="Reference beta*gamma", type=double, &end\n')
+        lines.insert(line1, '&parameter name=Step, description="Simulation step", type=long, &end\n')
+        
+        line2 = 18
+        lines.insert(line2, '28584M'+'\n')
+        lines.insert(line2, str(int(len(beam)))+'\n')
+        lines.insert(line2, str(int(len(beam)))+'\n')
+        lines.insert(line2, ' '+str(beam.charge())+'\n')
+        lines.insert(line2, ' '+str(energy2gamma(beam.energy()))+'\n')
+        lines.insert(line2, str(1)+'\n')
+        
+        with open(filename, 'w') as f:
+            f.write("".join(lines))
         
     # convert SDDS to binary
-    subprocess.call(CONFIG.elegant_exec + 'sddsconvert -binary ' + filename + ' -noWarnings', shell=True)
+    subprocess.run(CONFIG.elegant_exec + 'sddsconvert -binary ' + filename + ' -noWarnings', shell=True)
     
     # delete CSV file and temporary folder
-    shutil.rmtree(tmpfolder)
+    os.remove(tmpfile)
+    if make_new_tmpfolder:
+        shutil.rmtree(tmpfolder)
     
-    return filename
     
+def elegant_run(filename, beam0, inputbeamfile, outputbeamfile, envars={}, quiet=False, run_from_container=False, tmpfolder=None):
     
-def elegant_run(filename, beam0, beamfile, envars={}, quiet=False, run_from_container=False):
-    
-    # convert incoping beam object to temporary SDDS file
-    tmpfolder = CONFIG.temp_path + str(uuid.uuid4())
-    os.mkdir(tmpfolder)
-    tmpfile = tmpfolder + '/beam.bun'
-    elegant_write_beam(beam0, tmpfile)
-    envars['BEAM'] = tmpfile
-    
-    # set environment variables
-    for key in envars:
-        os.environ[key] = str(envars[key])
+    # convert incoming beam object to temporary SDDS file
+    elegant_write_beam(beam0, inputbeamfile, tmpfolder=tmpfolder)
 
     # run system command
     cmd = CONFIG.elegant_exec + 'elegant ' + filename + CONFIG.elegant_rpnflag
@@ -115,7 +116,7 @@ def elegant_run(filename, beam0, beamfile, envars={}, quiet=False, run_from_cont
         subprocess.call(cmd, shell=True, stdout=None)
         
     # convert SDDS output to beam object
-    beam = elegant_read_beam(beamfile)
+    beam = elegant_read_beam(outputbeamfile, tmpfolder=tmpfolder)
         
     # copy previous beam metadata
     beam.location = beam0.location
@@ -125,22 +126,21 @@ def elegant_run(filename, beam0, beamfile, envars={}, quiet=False, run_from_cont
     # reset previous macroparticle charge
     beam.copy_particle_charge(beam0)
     
-    # delete temporary bunch files and temporary folder
-    shutil.rmtree(tmpfolder)
-    
     return beam
 
 
-def elegant_apl_fieldmap2D(tau_lens, filename, lensdim_x=5e-3, lensdim_y=1e-3):
+def elegant_apl_fieldmap2D(tau_lens, filename, lensdim_x=5e-3, lensdim_y=1e-3, tmpfolder=None):
     
     # transverse dimensions
     xs = np.linspace(-lensdim_x, lensdim_x, 1001)
     ys = np.linspace(-lensdim_y, lensdim_y, 201)
     
     # create temporary CSV file and folder
-    tmpfolder = CONFIG.temp_path + str(uuid.uuid4())
-    os.mkdir(tmpfolder)
-    tmpfile = tmpfolder + '/map.csv'
+    make_new_tmpfolder = tmpfolder is None
+    if make_new_tmpfolder:
+        tmpfolder = CONFIG.temp_path + str(uuid.uuid4())
+        os.mkdir(tmpfolder)
+    tmpfile = tmpfolder + '/map_' + str(uuid.uuid4()) + '.csv'
     
     # create map
     Bmap = np.zeros((len(xs)*len(ys), 4));
@@ -165,6 +165,8 @@ def elegant_apl_fieldmap2D(tau_lens, filename, lensdim_x=5e-3, lensdim_y=1e-3):
                                                                         ' -columnData=name=By,type=double,unit=T', shell=True)
     
     # delete temporary CSV file and folder
-    shutil.rmtree(tmpfolder)
+    os.remove(tmpfile)
+    if make_new_tmpfolder:
+        shutil.rmtree(tmpfolder)
     
     return filename
